EntityDeleteConfirmModal: move to @labkey/components/entities
