+++ packages/components/src/index.ts
@@ -428,7 +428,6 @@ import {
 } from './internal/components/entities/constants';
 import { getUniqueIdColumnMetadata } from './internal/components/entities/utils';
 import { EntityInsertPanel } from './internal/components/entities/EntityInsertPanel';
-import { EntityDeleteConfirmModal } from './internal/components/entities/EntityDeleteConfirmModal';
 import { SampleTypeModel } from './internal/components/domainproperties/samples/models';
 
 import { EditableDetailPanel } from './public/QueryModel/EditableDetailPanel';
@@ -1055,7 +1054,6 @@ export {
     getParentTypeDataForLineage,
     getSelectedSampleIdsFromSelectionKey,
     EntityInsertPanel,
-    EntityDeleteConfirmModal,
     SampleTypeDataType,
     SamplePropertyDataType,
     DataClassDataType,
